Major Changes to the Bookworm PHP
Header and Footer are now in their own separate php files to be called whenever needed. Login in is half implemented. (Connection to database is set, and sign in button changes from "sign in" to "my account". Also login function works across each current page.  The customer id is stored as a session variable to be used in various other queries.  NOTICE: Cart item has slighlty shifted up,  as of right now unsure what causes this.